cloudfunctions/requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
fastjsonschema==2.14.4
get==2019.4.13
google-api-core==1.17.0
google-api-python-client==1.9.1
google-auth==1.16.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-error-reporting==0.34.0
google-cloud-firestore==1.7.0
google-cloud-logging==1.15.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.9
post==2019.4.13
protobuf==3.12.2
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
rsa==4.0
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.9
